integration/pyproject.toml
[project]
name = "trivyal-integration"
version = "0.1.0"
description = "Integration tests for Trivyal — requires Docker and Docker Compose"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9",
    "pytest-asyncio>=1.3",
    "pytest-docker>=3",
    "httpx>=0.28",
    "websockets>=14",
    "pynacl>=1.6",
    "requests>=2.32",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = ["integration: marks integration tests (require Docker)"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helpers"]
